Rename img_X_cls.py to img_X_cls_1000.py
