+++ benchmarker/modules/problems/res10ssd/data.py
@@ -1,1 +1,1 @@
-from benchmarker.util.data.synthetic.img_300_cls import get_data
+from benchmarker.util.data.synthetic.img_300_cls_1000 import get_data
