+++ benchmarker/modules/problems/vgg16/data.py
@@ -1,1 +1,1 @@
-from benchmarker.util.data.synthetic.img_224_cls import get_data
+from benchmarker.util.data.synthetic.img_224_cls_1000 import get_data
