+++ benchmarker/util/data/synthetic/img_224_cls_1000.py
