+++ benchmarker/util/data/synthetic/img_300_cls_1000.py
